requirements.txt
Flask>=1.0.2
psycopg2-binary>=2.8.3
requests>=2.6.0
beautifulsoup4>=4.7.1
selenium>=3.141.0
EZGmail>=0.0.8
